refactor: create classifier and regression in xgboost, move model and modeler to the same file
